added testrunner
